feat: funding sources map
