+++ src/modules/dashboard/components/ExpenditurePage/Stages/FormStages.tsx
@@ -24,7 +24,6 @@ interface Props {
   setFormValues?: React.Dispatch<React.SetStateAction<ValuesType | undefined>>;
   colony: Colony;
   handleCancelExpenditure: () => void;
-  formValues: ValuesType;
 }
 
 const FormStages = ({
@@ -34,7 +33,6 @@ const FormStages = ({
   setFormValues,
   colony,
   handleCancelExpenditure,
-  formValues,
 }: Props) => {
   const { values, handleSubmit, validateForm, resetForm, setTouched } =
     useFormikContext<ValuesType>() || {};
@@ -91,12 +89,7 @@ const FormStages = ({
   }, [activeState, handleSubmit, setTouched, validateForm, values]);
 
   if (values.expenditure === ExpenditureTypes.Streaming) {
-    return (
-      <StreamingStages
-        handleSaveDraft={handleSaveDraft}
-        formValues={formValues}
-      />
-    );
+    return <StreamingStages handleSaveDraft={handleSaveDraft} />;
   }
 
   return (
